Add footer and rename to ExamCalendar
