+++ www/src/js/views/timetable/ExamCalendar.scss
